+++ execution/src/main/java/com/multiassetoms/execution/model/OrderAcknowledgementException.java
@@ -1,0 +1,8 @@
+package com.multiassetoms.execution.model;
+
+public class OrderAcknowledgementException extends RuntimeException {
+
+    public OrderAcknowledgementException(String message) {
+        super(message);
+    }
+}