+++ execution/src/main/java/com/multiassetoms/execution/model/OrderFillException.java
@@ -1,0 +1,8 @@
+package com.multiassetoms.execution.model;
+
+public class OrderFillException extends RuntimeException {
+
+    public OrderFillException(String message) {
+        super(message);
+    }
+}